+++ src/types.ts
@@ -428,3 +428,22 @@ export interface IIndexArray<T> {
   };
   options: IndexOptions<T>;
 }
+
+export interface IGlobalOptions {
+  /** Typegoose Options */
+  options?: ICustomOptions;
+  /** Schema Options that should get applied to all models */
+  schemaOptions?: mongoose.SchemaOptions;
+  /**
+   * Global Options for general Typegoose
+   * (There are currently none)
+   */
+  globalOptions?: {
+    /**
+     * Use the new Enum code generating code
+     * -> this is for not breaking existing databases
+     * @default false
+     */
+    useNewEnum: boolean;
+  };
+}